+++ classes/post/wpdk-post.php
@@ -1,14 +1,4 @@
 <?php
-/// @cond private
-
-/*
- * [DRAFT]
- *
- * THE FOLLOWING CODE IS A DRAFT. FEEL FREE TO USE IT TO MAKE SOME EXPERIMENTS, BUT DO NOT USE IT IN ANY CASE IN
- * PRODUCTION ENVIRONMENT. ALL CLASSES AND RELATIVE METHODS BELOW CAN CHNAGE IN THE FUTURE RELEASES.
- *
- */
-
 /**
  * The WPDKPost class is a WordPress Post as object.
  *
@@ -27,12 +17,12 @@
  * @class              _WPDKPost
  * @author             =undo= <info@wpxtre.me>
  * @copyright          Copyright (C) 2012-2013 wpXtreme Inc. All Rights Reserved.
- * @date               2013-05-22
- * @version            1.0.0
+ * @date               2013-11-15
+ * @version            1.0.1
  *
  */
 
-class _WPDKPost {
+class WPDKPost extends WPDKObject {
 
   const COLUMN_COMMENT_COUNT         = 'comment_count';
   const COLUMN_COMMENT_STATUS        = 'comment_status';
@@ -57,6 +47,15 @@ class _WPDKPost {
   const COLUMN_POST_TITLE            = 'post_title';
   const COLUMN_POST_TYPE             = 'post_type';
   const COLUMN_TO_PING               = 'to_ping';
+
+  /**
+   * Override version
+   *
+   * @brief Version
+   *
+   * @var string $version
+   */
+  public $version = '1.0.1';
 
   /**
    * The post ID
@@ -809,7 +808,7 @@ class WPDKPostMeta {
    *
    * @brief An instance of _WPDKPost class
    *
-   * @var _WPDKPost $_post
+   * @var WPDKPost $_post
    */
   private $_post;
 
@@ -824,7 +823,7 @@ class WPDKPostMeta {
    */
   public function __construct( $post )
   {
-    $this->_post = new _WPDKPost( $post );
+    $this->_post = new WPDKPost( $post );
   }
 
   /**
@@ -891,5 +890,3 @@ class WPDKPostMeta {
     return null;
   }
 }
-
-/// @endcond