+++ CHANGELOG.md
@@ -20,6 +20,21 @@ For more information on SemVer, please visit http://semver.org/.
 
 ---
 
+## Version 1.4.1
+### 2013-11-??
+
+#### Enhancements
+
+* Improved layout of textarea in user profile
+* Minor performance improvements
+* Minor bugs fixes
+
+#### Under the hood
+
+* Removed old and deprecated `WPDKPost`
+* Renamed new `_WPDKPost` in `WPDKPost`
+
+
 ## Version 1.4.0
 ### 2013-11-14